requirements.txt
gdal == 2.1.3
numpy >= 1.13.2
matplotlib >= 2.0.2
pyshp >= 1.2.12
requirements-parser >= 0.1.0
